Move Ili2DbUtils to lib
